+++ src/main/java/backend/wal/auth/application/port/in/LoginRequestDto.java
@@ -2,7 +2,7 @@ package backend.wal.auth.application.port.in;
 
 import backend.wal.notification.application.port.in.InitFcmRequestDto;
 import backend.wal.user.domain.aggregate.SocialType;
-import backend.wal.user.application.port.dto.CreateUserDto;
+import backend.wal.user.application.port.in.CreateUserDto;
 import lombok.Getter;
 
 @Getter
