+++ src/main/java/backend/wal/user/adapter/ResignUserSchedulerAdapter.java
@@ -1,6 +1,6 @@
 package backend.wal.user.adapter;
 
-import backend.wal.user.application.port.ResignUserSchedulerPort;
+import backend.wal.user.application.port.out.ResignUserSchedulerPort;
 import org.springframework.stereotype.Component;
 
 import java.util.concurrent.Executors;
