--- src/main/java/backend/wal/user/application/port/RegisterResignInfoUseCase.java
@@ -1,8 +1,0 @@
-package backend.wal.user.application.port;
-
-import backend.wal.user.application.port.dto.RegisterResignRequestDto;
-
-public interface RegisterResignInfoUseCase {
-
-    void register(RegisterResignRequestDto requestDto);
-}
