--- src/main/java/backend/wal/user/application/port/RegisterUserUseCase.java
@@ -1,8 +1,0 @@
-package backend.wal.user.application.port;
-
-import backend.wal.user.application.port.dto.CreateUserDto;
-
-public interface RegisterUserUseCase {
-
-    Long signup(CreateUserDto createDto);
-}
