+++ src/main/java/backend/wal/user/application/port/in/ChangeUserInfoUseCase.java
@@ -1,4 +1,4 @@
-package backend.wal.user.application.port;
+package backend.wal.user.application.port.in;
 
 public interface ChangeUserInfoUseCase {
