+++ src/main/java/backend/wal/user/application/port/in/FindSocialUserUseCase.java
@@ -1,4 +1,4 @@
-package backend.wal.user.application.port;
+package backend.wal.user.application.port.in;
 
 import backend.wal.user.domain.aggregate.SocialType;
 import backend.wal.user.domain.aggregate.entity.User;
