+++ src/main/java/backend/wal/user/application/port/in/RegisterResignInfoUseCase.java
@@ -1,0 +1,6 @@
+package backend.wal.user.application.port.in;
+
+public interface RegisterResignInfoUseCase {
+
+    void register(RegisterResignRequestDto requestDto);
+}
