+++ src/main/java/backend/wal/user/application/port/in/RegisterUserUseCase.java
@@ -1,0 +1,6 @@
+package backend.wal.user.application.port.in;
+
+public interface RegisterUserUseCase {
+
+    Long signup(CreateUserDto createDto);
+}
