+++ src/main/java/backend/wal/user/application/service/RegisterUserService.java
@@ -1,7 +1,7 @@
 package backend.wal.user.application.service;
 
-import backend.wal.user.application.port.RegisterUserUseCase;
-import backend.wal.user.application.port.dto.CreateUserDto;
+import backend.wal.user.application.port.in.RegisterUserUseCase;
+import backend.wal.user.application.port.in.CreateUserDto;
 import backend.wal.user.domain.aggregate.entity.User;
 import backend.wal.user.domain.repository.UserRepository;
 import backend.wal.support.annotation.AppService;
